+++ lib/src/typedefs.dart
@@ -40,6 +40,9 @@ typedef Stemmer = String Function(Term term);
 /// A language-specific function that returns the lemma of [term].
 typedef Lemmatizer = String Function(Term term);
 
+/// A language-specific function that generates n-grams from text.
+typedef NGrammer = List<String> Function(String text, NGramRange range);
+
 /// A language-specific function that returns the number of syllables in a
 /// string after stripping out all white-space and punctuation.
 ///